chore: move user roles test file
